Game Over Screen, Score, and High Score
Sends the player to a Game Over screen on death, from which the game can be restarted. The player now gains a point each time a monster is killed, which is displayed in the top right corner. Your score and high score are both displayed on the game over screen, and your high score is saved across playthroughs